+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - upgrade path for items
 - emojified Ability and Item properties
 - enabled slash commands
+- `invite`, `help` and `repo` commands
 
 ## [v0.2.0] - 2026-04-19
 - rework Item and Ability data structures